made UIThread method abstract, made fields that are not called by inheritors private
