+++ src/EditorFeatures/Core.Wpf/Adornments/AdornmentManager.cs
@@ -24,9 +24,15 @@ namespace Microsoft.CodeAnalysis.Editor.Implementation.Adornments
     /// </summary>
     internal abstract class AdornmentManager<T> where T : GraphicsTag
     {
-        protected readonly object _invalidatedSpansLock = new object();
+        private readonly object _invalidatedSpansLock = new object();
 
-        protected IThreadingContext ThreadingContext { get; }
+        private readonly IThreadingContext _threadingContext;
+
+        /// <summary>Notification system about operations we do</summary>
+        private readonly IAsynchronousOperationListener _asyncListener;
+
+        /// <summary>Spans that are invalidated, and need to be removed from the layer..</summary>
+        private List<IMappingSpan> _invalidatedSpans;
 
         /// <summary>View that created us.</summary>
         protected IWpfTextView TextView { get; }
@@ -37,11 +43,15 @@ namespace Microsoft.CodeAnalysis.Editor.Implementation.Adornments
         /// <summary>Aggregator that tells us where to draw.</summary>
         protected ITagAggregator<T> TagAggregator { get; }
 
-        /// <summary>Notification system about operations we do</summary>
-        protected IAsynchronousOperationListener AsyncListener { get; }
-
-        /// <summary>Spans that are invalidated, and need to be removed from the layer..</summary>
-        protected List<IMappingSpan> _invalidatedSpans;
+        /// <summary>
+        /// MUST BE CALLED ON UI THREAD!!!!   This method touches WPF.
+        /// 
+        /// This is where we apply visuals to the text. 
+        /// 
+        /// It happens when another region of the view becomes visible or there is a change in tags.
+        /// For us the end result is the same - get tags from tagger and update visuals correspondingly.
+        /// </summary>        
+        protected abstract void UpdateSpans_CallOnlyOnUIThread(NormalizedSnapshotSpanCollection changedSpanCollection, bool removeOldTags);
 
         internal AdornmentManager(
             IThreadingContext threadingContext,
@@ -56,11 +66,11 @@ namespace Microsoft.CodeAnalysis.Editor.Implementation.Adornments
             Contract.ThrowIfNull(adornmentLayerName);
             Contract.ThrowIfNull(asyncListener);
 
-            ThreadingContext = threadingContext;
+            _threadingContext = threadingContext;
             TextView = textView;
             AdornmentLayer = textView.GetAdornmentLayer(adornmentLayerName);
             textView.LayoutChanged += OnLayoutChanged;
-            AsyncListener = asyncListener;
+            _asyncListener = asyncListener;
 
             // If we are not on the UI thread, we are at race with Close, but we should be on UI thread
             Contract.ThrowIfFalse(textView.VisualElement.Dispatcher.CheckAccess());
@@ -91,7 +101,7 @@ namespace Microsoft.CodeAnalysis.Editor.Implementation.Adornments
         private void OnLayoutChanged(object sender, TextViewLayoutChangedEventArgs e)
         {
             using (Logger.LogBlock(FunctionId.Tagger_AdornmentManager_OnLayoutChanged, CancellationToken.None))
-            using (AsyncListener.BeginAsyncOperation(GetType() + ".OnLayoutChanged"))
+            using (_asyncListener.BeginAsyncOperation(GetType() + ".OnLayoutChanged"))
             {
                 // Make sure we're on the UI thread.
                 Contract.ThrowIfFalse(TextView.VisualElement.Dispatcher.CheckAccess());
@@ -137,7 +147,7 @@ namespace Microsoft.CodeAnalysis.Editor.Implementation.Adornments
         /// </summary>
         private void OnTagsChanged(object sender, TagsChangedEventArgs e)
         {
-            using (AsyncListener.BeginAsyncOperation(GetType().Name + ".OnTagsChanged.1"))
+            using (_asyncListener.BeginAsyncOperation(GetType().Name + ".OnTagsChanged.1"))
             {
                 var changedSpan = e.Span;
 
@@ -166,11 +176,11 @@ namespace Microsoft.CodeAnalysis.Editor.Implementation.Adornments
                 if (needToScheduleUpdate)
                 {
                     // schedule an update
-                    ThreadingContext.JoinableTaskFactory.WithPriority(TextView.VisualElement.Dispatcher, DispatcherPriority.Render).RunAsync(async () =>
+                    _threadingContext.JoinableTaskFactory.WithPriority(TextView.VisualElement.Dispatcher, DispatcherPriority.Render).RunAsync(async () =>
                     {
-                        using (AsyncListener.BeginAsyncOperation(GetType() + ".OnTagsChanged.2"))
+                        using (_asyncListener.BeginAsyncOperation(GetType() + ".OnTagsChanged.2"))
                         {
-                            await ThreadingContext.JoinableTaskFactory.SwitchToMainThreadAsync(alwaysYield: true);
+                            await _threadingContext.JoinableTaskFactory.SwitchToMainThreadAsync(alwaysYield: true);
                             UpdateInvalidSpans();
                         }
                     });
@@ -185,7 +195,7 @@ namespace Microsoft.CodeAnalysis.Editor.Implementation.Adornments
         /// </summary>
         private void UpdateInvalidSpans()
         {
-            using (AsyncListener.BeginAsyncOperation(GetType().Name + ".UpdateInvalidSpans.1"))
+            using (_asyncListener.BeginAsyncOperation(GetType().Name + ".UpdateInvalidSpans.1"))
             using (Logger.LogBlock(FunctionId.Tagger_AdornmentManager_UpdateInvalidSpans, CancellationToken.None))
             {
                 // this method should only run on UI thread as we do WPF here.
@@ -208,99 +218,6 @@ namespace Microsoft.CodeAnalysis.Editor.Implementation.Adornments
                     var viewSnapshot = TextView.TextSnapshot;
                     var invalidatedNormalized = TranslateAndNormalize(invalidated, viewSnapshot);
                     UpdateSpans_CallOnlyOnUIThread(invalidatedNormalized, removeOldTags: true);
-                }
-            }
-        }
-
-        /// <summary>
-        /// MUST BE CALLED ON UI THREAD!!!!   This method touches WPF.
-        /// 
-        /// This is where we apply visuals to the text. 
-        /// 
-        /// It happens when another region of the view becomes visible or there is a change in tags.
-        /// For us the end result is the same - get tags from tagger and update visuals correspondingly.
-        /// </summary>        
-        protected virtual void UpdateSpans_CallOnlyOnUIThread(NormalizedSnapshotSpanCollection changedSpanCollection, bool removeOldTags)
-        {
-            Contract.ThrowIfNull(changedSpanCollection);
-
-            // this method should only run on UI thread as we do WPF here.
-            Contract.ThrowIfFalse(TextView.VisualElement.Dispatcher.CheckAccess());
-
-            var viewSnapshot = TextView.TextSnapshot;
-            var visualSnapshot = TextView.VisualSnapshot;
-
-            var viewLines = TextView.TextViewLines;
-            if (viewLines == null || viewLines.Count == 0)
-            {
-                return; // nothing to draw on
-            }
-
-            // removing is a separate pass from adding so that new stuff is not removed.
-            if (removeOldTags)
-            {
-                foreach (var changedSpan in changedSpanCollection)
-                {
-                    // is there any effect on the view?
-                    if (viewLines.IntersectsBufferSpan(changedSpan))
-                    {
-                        AdornmentLayer.RemoveAdornmentsByVisualSpan(changedSpan);
-                    }
-                }
-            }
-
-            foreach (var changedSpan in changedSpanCollection)
-            {
-                // is there any effect on the view?
-                if (!viewLines.IntersectsBufferSpan(changedSpan))
-                {
-                    continue;
-                }
-
-                var tagSpans = TagAggregator.GetTags(changedSpan);
-                foreach (var tagMappingSpan in tagSpans)
-                {
-                    // We don't want to draw line separators if they would intersect a collapsed outlining
-                    // region.  So we test if we can map the start of the line separator up to our visual 
-                    // snapshot. If we can't, then we just skip it.
-                    var point = tagMappingSpan.Span.Start.GetPoint(changedSpan.Snapshot, PositionAffinity.Predecessor);
-                    if (point == null)
-                    {
-                        continue;
-                    }
-
-                    var mappedPoint = TextView.BufferGraph.MapUpToSnapshot(
-                        point.Value, PointTrackingMode.Negative, PositionAffinity.Predecessor, TextView.VisualSnapshot);
-                    if (mappedPoint == null)
-                    {
-                        continue;
-                    }
-
-                    if (!TryMapToSingleSnapshotSpan(tagMappingSpan.Span, viewSnapshot, out var span))
-                    {
-                        continue;
-                    }
-
-                    if (!viewLines.IntersectsBufferSpan(span))
-                    {
-                        // span is outside of the view so we will not get geometry for it, but may 
-                        // spent a lot of time trying.
-                        continue;
-                    }
-
-                    // add the visual to the adornment layer.
-                    var geometry = viewLines.GetMarkerGeometry(span);
-                    if (geometry != null)
-                    {
-                        var tag = tagMappingSpan.Tag;
-                        var graphicsResult = tag.GetGraphics(TextView, geometry);
-                        AdornmentLayer.AddAdornment(
-                            behavior: AdornmentPositioningBehavior.TextRelative,
-                            visualSpan: span,
-                            tag: tag,
-                            adornment: graphicsResult.VisualElement,
-                            removedCallback: delegate { graphicsResult.Dispose(); });
-                    }
                 }
             }
         }
